Screenshots are working now.
The example from the pluralsight course had a fundamental problem.
The API for getting screen info is no longer callable from a Renderer
process. To address that, I get the screen info from the Main process,
and use the IPC tools to send the screen size info to the Renderer
process.